add start menu

+++ models/ppo.py
@@ -5,65 +5,31 @@ from stable_baselines3 import PPO
 from stable_baselines3.common.env_util import make_vec_env
 from stable_baselines3.common.evaluation import evaluate_policy
 from stable_baselines3.common.callbacks import EvalCallback
-from render.user_interface import show_input_window, show_test_prompt_window, quit_pygame
-from envs.fire_env import FireEnv
 
-summary_shown = False
+from models.test_model import test_model
+from render.user_interface import show_input_window, show_test_prompt_window, show_start_window
+from envs.fire_env import FireEnv
+from utils.logging_files import log_csv, tensorboard_log_dir
 
 
 def run():
-    global summary_shown
-    log_csv = "./logs/logs.csv"
-    if os.path.exists(log_csv):
-        os.remove(log_csv)
-    with open(log_csv, mode='w', newline='') as file:
-        writer = csv.writer(file)
-        writer.writerow(["Episode", "Step", "Battery1", "Battery2", "Battery3",
-                         "Extinguishers1", "Extinguishers2", "Extinguishers3",
-                         "Fires Left", "Reward", "Action1", "Action2", "Action3"])
+    if show_start_window():
 
-    fire_count, obstacle_count = show_input_window()
-    logging.info("Starting model training...")
-    model = train_and_evaluate(fire_count, obstacle_count)
-    logging.info("Training completed!")
+        if os.path.exists(log_csv):
+            os.remove(log_csv)
+        with open(log_csv, mode='w', newline='') as file:
+            writer = csv.writer(file)
+            writer.writerow(["Episode", "Step", "Battery1", "Battery2", "Battery3",
+                             "Extinguishers1", "Extinguishers2", "Extinguishers3",
+                             "Fires Left", "Reward", "Action1", "Action2", "Action3"])
 
-    if show_test_prompt_window():
-        test_env = FireEnv(fire_count=fire_count, obstacle_count=obstacle_count, render_mode="human")
-        logging.info("Starting model testing...")
-        for episode in range(1, 5):
-            obs, _ = test_env.reset()
-            total_reward = 0
-            while True:
-                actions, _ = model.predict(obs)  # Единое действие для всех агентов
-                obs, reward, terminated, truncated, info = test_env.step(actions)
-                total_reward += reward
-                with open(log_csv, mode='a', newline='') as file:
-                    writer = csv.writer(file)
-                    writer.writerow([episode, test_env.iteration_count] +
-                                    test_env.battery_levels + test_env.extinguisher_counts +
-                                    [len(test_env.fires), reward] + list(actions))
-                logging.info(f"Episode = {episode}, Step {test_env.iteration_count + 1}: "
-                             f"Reward = {reward}, Total Reward = {total_reward}, "
-                             f"Fires Left: {len(test_env.fires)}")
-                test_env.render()
-                if terminated or truncated:
-                    if len(test_env.fires) == 0:
-                        print(f"Testing completed at step {test_env.iteration_count}: "
-                              f"All fires extinguished! Total Reward: {total_reward}")
-                    elif test_env.iteration_count + 1 >= test_env.max_steps:
-                        print(f"Testing completed at step {test_env.iteration_count}: "
-                              f"Step limit reached. Total Reward: {total_reward}")
-                    else:
-                        print(f"Testing completed at step {test_env.iteration_count} "
-                              f"for unknown reason. Total Reward: {total_reward}")
-                    if not summary_shown:
-                        test_env.close()
-                        summary_shown = True
-                    break
-        if not summary_shown:
-            test_env.close()
-            summary_shown = True
-    quit_pygame()
+        fire_count, obstacle_count = show_input_window()
+        logging.info("Starting model training...")
+        model = train_and_evaluate(fire_count, obstacle_count)
+        logging.info("Training completed!")
+
+        if show_test_prompt_window():
+            test_model(model, fire_count, obstacle_count)
 
 
 def train_and_evaluate(fire_count, obstacle_count):
@@ -71,8 +37,7 @@ def train_and_evaluate(fire_count, obstacle_count):
         return FireEnv(fire_count=fire_count, obstacle_count=obstacle_count, render_mode=None)
 
     vec_env = make_vec_env(make_env, n_envs=1)
-    log_dir = "./logs/ppo_tensorboard/"
-    os.makedirs(log_dir, exist_ok=True)
+    os.makedirs(tensorboard_log_dir, exist_ok=True)
     model = PPO(
         "MlpPolicy",
         vec_env,
@@ -86,12 +51,12 @@ def train_and_evaluate(fire_count, obstacle_count):
         clip_range=0.2,
         clip_range_vf=0.2,
         ent_coef=0.01,
-        tensorboard_log=log_dir
+        tensorboard_log=tensorboard_log_dir
     )
     eval_callback = EvalCallback(
         vec_env,
         best_model_save_path="./data/best_model/",
-        log_path=log_dir,
+        log_path=tensorboard_log_dir,
         eval_freq=1000,
         render=False
     )
